src/requirements.txt
google-cloud
google-cloud-bigquery
google-cloud-storage
google-cloud-logging
google-cloud-pubsub